completed technical & marketing testing

+++ verp_staffing/employee/doctype/employee/employee.py
@@ -18,6 +18,8 @@ class Employee(Document):
         self.name = generate_name_series("Employee", name)
 
     def on_update(self):
+        self.validate_employee_edit_permission()
+
         # Clear cached visible employee names for the user whenever an Employee is updated,
         # to ensure any changes are reflected in reports immediately
         frappe.cache().delete_key("Visible_Employee_Names")
@@ -25,7 +27,7 @@ class Employee(Document):
         frappe.cache().delete_key(f"opportunity_user::{self.user}")
         frappe.cache().delete_key(f"employee_from_user::{self.user}")
         frappe.cache().delete_key(f"user_departments::{self.name}")
-        
+
     def on_trash(self):
         validate_employee_delete(self)
         frappe.cache().delete_key("Visible_Employee_Names")
@@ -35,6 +37,7 @@ class Employee(Document):
         frappe.cache().delete_key(f"user_departments::{self.name}")
 
     def validate(self):
+        self.validate_employee_edit_permission()
 
         rows = self.employee_assignment_details_table or []
 
@@ -193,7 +196,7 @@ def get_users_not_linked_to_employee(
         SELECT
             u.name
         FROM `tabUser` u
-        WHERE {' AND '.join(conditions)}
+        WHERE {" AND ".join(conditions)}
         ORDER BY u.name ASC
         LIMIT %s OFFSET %s
         """,
@@ -316,6 +319,7 @@ def get_employee_from_user(user):
 
     return employee
 
+
 @frappe.whitelist()
 def get_user_departments(user=None):
 
@@ -347,7 +351,6 @@ def get_user_departments(user=None):
     )
 
     return departments
-
 
 
 def validate_employee_assignment_hierarchy(doc):
@@ -559,7 +562,7 @@ def validate_employee_assignment_hierarchy(doc):
         #             current_role,
         #             ", ".join(sorted(allowed_parent_roles)),
         #         )
-            # )
+        # )
 
         assigned_roles = employee_role_map.get(
             assigned_to,
